requirements.txt
flask
requests
pycryptodome
protobuf==4.25.3
gunicorn
